added total time watched for all videos
